requirements.txt
requests~=2.26.0
Flask~=2.0.2
WTForms~=3.0.1
Flask-WTF~=0.15.1
gunicorn~=20.1.0
python-dotenv~=0.19.2